feat: implement message extraction and transformation for tool execution history
